Literature to understand Soncini
The key papers are the Griffith ones, particularly parts 1 and 3. These apply the theory detailed in the Fano_Racah textbook. 

The Qiu-Ceulemans paper contains coefficients which are useful for determining the matrix representations of the operators.

The stos.pdf/tex files are my own notes, these are very rough, but might be useful as a quick introduction.